// src/app/models/pokemon.ts
export interface pokemon {
    id: number;
    abilities: string;
    base_experience: number;
    forms: string;
    game_indices: number;
    height: number;
    held_items: string;
    is_default: boolean;
    location_area_encounters: string;
    moves: string;
    order: number;
    name: string;
    species: string;
    sprites: string;
    stats: number;
    types: string;
    weight: number;
  }